Move some image functions into separate file
